预测功能完成，待测试

+++ cpf/cpf-dump/src/main/java/com/cpf/mysql/dao/PO/ModelPO.java
@@ -17,7 +17,7 @@ public class ModelPO {
     /**
      * 模型占决策的比重
      */
-    private Long weight = 0L;
+    private Double weight = 0D;
     /**
      * 模型参数配置
      */
