+++ cpf/cpf-dump/src/main/java/com/cpf/mysql/manager/DO/ModelDO.java
@@ -14,7 +14,7 @@ public class ModelDO {
     /**
      * 模型占决策的比重
      */
-    private Long weight = 0L;
+    private Double weight =0D;
     /**
      * 模型参数配置
      */
